Centralize Codex CLI discovery
