+++ rust/src/cli/tty_runner.rs
@@ -191,7 +191,7 @@ impl TtyCommandRunner {
     pub fn which(tool: &str) -> Option<PathBuf> {
         // Check for specific tool overrides
         if tool == "codex"
-            && let Some(path) = Self::locate_codex_binary()
+            && let Some(path) = crate::codex_cli::locate_codex_binary()
         {
             return Some(path);
         }
@@ -203,40 +203,6 @@ impl TtyCommandRunner {
 
         // Use `where` on Windows (equivalent to `which` on Unix)
         Self::run_where(tool)
-    }
-
-    /// Locate the Codex binary
-    fn locate_codex_binary() -> Option<PathBuf> {
-        // Check environment override
-        if let Ok(path) = std::env::var("CODEX_BINARY") {
-            let path = PathBuf::from(path);
-            if path.exists() {
-                return Some(path);
-            }
-        }
-
-        // Check common Windows locations
-        let candidates = [
-            // npm global install locations
-            dirs::data_local_dir().map(|d| d.join("npm").join("codex.cmd")),
-            dirs::home_dir().map(|h| {
-                h.join("AppData")
-                    .join("Roaming")
-                    .join("npm")
-                    .join("codex.cmd")
-            }),
-            // Bun install
-            dirs::home_dir().map(|h| h.join(".bun").join("bin").join("codex.exe")),
-        ];
-
-        for candidate in candidates.into_iter().flatten() {
-            if candidate.exists() {
-                return Some(candidate);
-            }
-        }
-
-        // Fall back to PATH search
-        Self::run_where("codex")
     }
 
     /// Locate the Claude binary
