+++ rust/src/lib.rs
@@ -7,6 +7,7 @@ pub mod agent_sessions;
 pub mod browser;
 pub mod cli;
 pub mod codex_accounts;
+pub(crate) mod codex_cli;
 pub mod codex_workspaces;
 pub mod core;
 pub mod cost_scanner;
